requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
argcomplete=1.12.3=pyhd3eb1b0_0
argon2-cffi=20.1.0=py37h27cfd23_1
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.5.7=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bottleneck=1.3.2=py37heb32a55_1
ca-certificates=2021.9.30=h06a4308_1
certifi=2021.10.8=py37h06a4308_0
cffi=1.14.6=py37h400218f_0
cycler=0.10.0=py37_0
dbus=1.13.18=hb2f20db_0
debugpy=1.4.1=py37h295c915_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
importlib-metadata=4.8.1=py37h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2020.2=254
ipykernel=6.2.0=py37h06a4308_1
ipython=7.27.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py37h06a4308_1
jinja2=3.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_core=4.7.1=py37h06a4308_0
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.34.0=py37h269e1b5_4
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py37h27cfd23_0
matplotlib=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
nbconvert=5.6.1=py37_1
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.3=py37h06a4308_0
numba=0.51.2=py37h04863e7_1
numexpr=2.7.3=py37hb2eb853_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.2=py37h8c16a72_0
pandoc=2.12=h06a4308_0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickle5=0.0.11=py37h27cfd23_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py37h5aabda8_0
pip=21.2.2=py37h06a4308_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyhca03da5_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pyzmq=22.2.1=py37h295c915_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
rope=0.19.0=pyhd3eb1b0_0
scipy=1.6.2=py37h91f5cce_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tbb=2021.3.0=hd09550d_0
terminado=0.9.4=py37h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py37h27cfd23_0
tqdm=4.62.2=pyhd3eb1b0_1
traitlets=5.1.0=pyhd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
